+++ chrome/background.js
@@ -25,19 +25,18 @@ chrome.browserAction.onClicked.addListener(
 );
 
 /**
- * Occurs when user clicks "insert" button.
  * Popup cannot access DOM or talk to content script directly, so this
- * listener runs in the background to relay the "insert_text" command from one to the other
+ * listener runs in the background to relay ANY (1.2) command from popup.js
+ * to contentscript.js
  */
 chrome.extension.onMessage.addListener(
 	function(request, sender, sendResponse) {
-		var word = request.pass;
-		sendResponse({confirmation: word});
 		
 		// relay message on to content script
 		chrome.tabs.query({active: true, currentWindow: true}, function(tabs){
-			var obj = {action: "insert_text", pass: word};
-			chrome.tabs.sendMessage(tabs[0].id, obj, function(response) {});
+			chrome.tabs.sendMessage(tabs[0].id, request, function(response) {
+				//sendResponse({'request': request, 'response': response});
+			});
 		});
 	}
 );