+++ chrome/popup.js
@@ -31,23 +31,34 @@ var pass = {
     var site = document.getElementById("hash_site").value;
 		var plaintext = site + phrase;
 		
-		// 1.1 extra special security: encourage passphrase to be longer than 10 characters
-		phrase_len = phrase.length;
-		bg = "";
-		if (phrase_len < 10) {
-			bg = "red";
+		// 1.2 don't do any work if the field is empty
+		if (phrase.length > 0) {
+			// show show/hide icon
+			document.getElementById('showhide').style.display = 'inline';
+		
+			// 1.1 extra special security: encourage passphrase to be
+			// longer than 10 characters
+			phrase_len = phrase.length;
+			bg = "";
+			if (phrase_len < 10) {
+				bg = "red";
+			} else {
+				bg = "black";
+			}
+			document.getElementById('hash_result').style.color = bg;
+			
+			// hash away
+			var ciphertext = SHA256(plaintext);
+			ciphertext = this.HexToBase64(ciphertext);
+			ciphertext = ciphertext.substring(0,16);
+			
+			// display result
+			document.getElementById('hash_result').innerHTML = ciphertext;
 		} else {
-			bg = "black";
+			// hide show/hide icon, clear result
+			document.getElementById('showhide').style.display = 'none';
+			document.getElementById('hash_result').innerHTML = '';
 		}
-		document.getElementById("hash_result").style.color = bg;
-		
-		// hash away
-		var ciphertext = SHA256(plaintext);
-		ciphertext = this.HexToBase64(ciphertext);
-		ciphertext = ciphertext.substring(0,16);
-		
-		// display result
-		document.getElementById("hash_result").innerHTML = ciphertext;
   },
   
 	/**
@@ -55,14 +66,60 @@ var pass = {
 	 * and initialize all inputs
 	 */
   insert: function(word) {
-		//console.log('sending password "'+word+'" to background...');
-		chrome.extension.sendMessage({pass:word}, function(response){
-			//console.log('received: '+response.confirmation);
-			
-			// clear inputs
-			document.getElementById("hash_phrase").value = "";
-			document.getElementById("hash_result").innerHTML = "";
-		});
+		var msg = { action: 'insert_text', pass: word };
+		if (word.length > 0) {
+			chrome.extension.sendMessage(msg, function(response){});
+		}
+	},
+
+	/**
+	* short utility functions
+	*/
+	toggle: function() {
+		var type = document.getElementById('hash_phrase').type;
+		
+		var show = (type == 'password');
+		var fieldtype = show ? 'text': 'password';
+		var imgsrc    = show ? 'hide20.png': 'show20.png';
+		var imgtitle  = show ? 'Hide': 'Show';
+		
+		document.getElementById('hash_phrase').type = fieldtype;
+		document.getElementById('showhide').src = imgsrc;
+		document.getElementById('showhide').title = imgtitle+' Passphrase Alt+S';
+	},
+	init: function(which) {
+		this.toggle('hide');
+		document.getElementById('hash_result').innerHTML = '';
+		document.getElementById('hash_phrase').value = '';
+		document.getElementById('showhide').style.display = 'none';
+		
+		if (which == 'visible') {
+			document.addEventListener('keydown', pass.listen);
+			document.getElementById('hash_phrase').focus();
+		} else {
+			// 1.2 remove listeners!
+			document.removeEventListener('keydown', pass.listen);
+		}
+	},
+	
+	/*
+	* key listener
+	* 1.2 make sure to disable on qtip hide
+	*/
+	listen: function(e) {
+		var key = e.keyCode;
+		if (key == 13) { /* enter */
+			var word = document.getElementById("hash_result").innerHTML;
+			pass.insert(word);
+		} else if (e.altKey & (key == 83)) { /* alt+s */
+			pass.toggle();
+		} else if (key == 27) { /* esc */
+			// 1.2 tell content script to close qtip
+			chrome.extension.sendMessage(
+				{ action: 'close' },
+				function(response){}
+			);
+		}
 	}
 };
 
@@ -81,7 +138,7 @@ document.addEventListener('DOMContentLoaded', function() {
 	// make the result text clickable, click to insert, again, saving a button
 	var hash_insert = document.getElementById('hash_insert');
 	hash_insert.addEventListener('click', function() {
-		var word = document.getElementById("hash_result").innerHTML;
+		var word = document.getElementById('hash_result').innerHTML;
 		pass.insert(word);
 	});
 	
@@ -103,35 +160,36 @@ document.addEventListener('DOMContentLoaded', function() {
 		// https://www.google.com/s2/favicons?domain=www.[] misses github
 		// https://plus.google.com/_/favicon?domain=www.[] misses twitter
 		// https://plus.google.com/_/favicon?domain=[] misses dropbox
-		var iconsrc = icon;
-		if (iconsrc == "") {
-			iconsrc = "http://g.etfv.co/http://www."+domain;
+		if (icon == "") {
+			icon = "http://g.etfv.co/http://www."+domain;
 		}
-		hash_site.style.backgroundImage = "url('"+iconsrc+"')";
+		hash_site.style.backgroundImage = "url('"+icon+"')";
 		hash_site.style.backgroundRepeat = "no-repeat";
 		hash_site.style.backgroundSize = "contain";
 		hash_site.style.paddingLeft = "20px";
 	});
+
+	/**
+	* Show/hide password toggle
+	*/
+	document.getElementById('showhide').addEventListener('click', function() {
+		pass.toggle();
+	});
 	
-	document.getElementById("hash_phrase").focus();
+	// autofocus passphrase
+	pass.init();
 });
 
 /*
- * insert on enter
- */
-document.addEventListener('keydown', function(e) {
-	var key = e.keyCode;
-	if (key == 13) {
-		var word = document.getElementById("hash_result").innerHTML;
-		pass.insert(word);
-	}
-});
-
-/*
- * If user enters, then exits, then enters qtip, refocus passphrase
+ * 1.1 If user enters, then exits, then enters qtip, refocus passphrase
  * (ondomloaded is catching this, but onshow can't reach inside the iframe
  *  to focus because of security restrictions -- use post message)
  */
 window.addEventListener('message',function(e) {
-	document.getElementById('hash_phrase').focus();
+	if (e.data.visible) { /* on qtip show */
+		pass.init('visible');
+	}
+	if (e.data.hidden) { /* on qtip hide */
+		pass.init('hidden');
+	}
 });